PORT-010: Retrieving images from Instagram API (#37)
## 👨🏽‍💻 Description
These changes aim to perform the integration with the Instagram API in
order to get the pictures that will be shown in the Homepage Hero
Slider. The integration is performed in the back-end with the gem
`InstagramBasicDisplay`. I set up a Facebook App in the Facebook
Developers website to get access to my Facebook and Instagram accounts
and pictures.

The URL and caption of the images are included in the JSON for the
homepage coming from `Curriculum` model. I set up a cache of `1` day to
renew the file with the results.

Then, in the front-end side the `HeroSlider` component is populated with
the images and creates a `Carousel` with them.

The changes also include several improvements on the styling and the
navigation.

These changes needs to be deployed after populate the secrets
provisioned by following Pull request in Lorentz:
* [LRNT-012: Adding Instagram access to get
pictures](https://github.com/zatarain/lorentz/pull/21)


## ✅ Testing
The changes include unit testing for the impacted components on both
(front-end and back-end). Which were mainly:
* Back-end - Model - `Curriculum`
* Front-end - Component - `HeroSlider`
* Front-end - Component/slice - `NavigationBar`

Changes were also tested manually on local environment and also in
`development` and `staging`.

## 🖼️ Evidence
Following screenshot shows the `staging` website:

![image](https://github.com/zatarain/portfolio/assets/539783/82043ba9-1527-4624-8733-8028c66611fe)

And the next one shows what the website looks like in my phone:
![WhatsApp Image 2023-07-06 at 05 07
41](https://github.com/zatarain/portfolio/assets/539783/0ef6712d-5f02-41b4-abf7-0ebdc4a2e022)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>