PORT-012: Adding database configuration (#48)
## 👨🏽‍💻 Description

These changes include the database configuration and the creation of the
empty database.

They also in include an improvement for the workflow in the back-end and
the Markdown template for the pull requests.

## 🎫 Related pull requests

This pull request needs to be deployed after following: 
* zatarain/lorentz#23

## ✅ Testing

These changes were merged to development and staging branches in order
to deploy them to those environments. I verified the the database was
created there.

Since the database instance ins staging is not publicly reachable I
connected to the API service task in AWS ECS and then connect to
database instance from there.

## 🖼️ Evidence

I connected to `development` database instance and verified that the
database was created. Following screenshot shows the database in
`pgAdmin`:

![image](https://github.com/zatarain/portfolio/assets/539783/fbe7034d-c52e-44c9-8706-68f64d18e923)

Following screenshot shoes the databases in `staging`:

![image](https://github.com/zatarain/portfolio/assets/539783/eaabc762-383c-4779-b317-905ec27d5d08)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>