PORT-014: Installing Storybook to test components in isolated way (#80)
## 👨🏽‍💻 Description

This changes add Storybook to the project in order to be able to test
isolated react components. It also add the first story for `Markdown`
component. Also, the documentation and pull request template were
updated.

Additionally, I updated the pipelines for GitHub Actions to run them
always on push for both (back-end and front-end).

## ✅ Testing

 * [x] I ran Storybook locally with `npm run storybook`
 * [x] Adding the first story for `Markdown` component

## 🖼️ Evidence

Following screenshot shows how the Storybook UI is running locally:


![image](https://github.com/zatarain/portfolio/assets/539783/32b1c0ff-a8df-40ae-bf66-803eece3a1c6)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>